backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
telethon==1.34.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
cryptography==42.0.0
python-socketio==5.11.0
